+++ package.json
@@ -28,20 +28,20 @@
     "@babel/preset-flow": "^7.25.9",
     "@babel/preset-react": "^7.23.3",
     "@babel/preset-typescript": "^7.23.3",
-    "@chromatic-com/storybook": "^3.2.4",
-    "@storybook/addon-a11y": "^8.6.0",
-    "@storybook/addon-actions": "^8.6.0",
-    "@storybook/addon-essentials": "^8.6.0",
-    "@storybook/addon-interactions": "^8.6.0",
-    "@storybook/addon-links": "^8.6.0",
-    "@storybook/addon-mdx-gfm": "^8.6.0",
+    "@chromatic-com/storybook": "^3.2.6",
+    "@storybook/addon-a11y": "^8.6.6",
+    "@storybook/addon-actions": "^8.6.6",
+    "@storybook/addon-essentials": "^8.6.6",
+    "@storybook/addon-interactions": "^8.6.6",
+    "@storybook/addon-links": "^8.6.6",
+    "@storybook/addon-mdx-gfm": "^8.6.6",
     "@storybook/addon-webpack5-compiler-babel": "^3.0.5",
-    "@storybook/manager-api": "^8.6.0",
-    "@storybook/preview-api": "^8.6.0",
-    "@storybook/react": "^8.6.0",
-    "@storybook/react-webpack5": "^8.6.0",
+    "@storybook/manager-api": "^8.6.6",
+    "@storybook/preview-api": "^8.6.6",
+    "@storybook/react": "^8.6.6",
+    "@storybook/react-webpack5": "^8.6.6",
     "@storybook/storybook-deployer": "^2.8.16",
-    "@storybook/test": "^8.6.0",
+    "@storybook/test": "^8.6.6",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
     "@typescript-eslint/parser": "^6.15.0",
     "babel-loader": "^9.1.3",
@@ -53,10 +53,10 @@
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-react-compiler": "^19.0.0-beta-40c6c23-20250301",
     "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-plugin-storybook": "^0.11.3",
+    "eslint-plugin-storybook": "^0.11.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "storybook": "^8.6.0",
+    "storybook": "^8.6.6",
     "trim": "^1.0.1",
     "typescript": "^5.2.0",
     "webpack": "^5.94.0",
