+++ packages/css/src/formElements/select/select.css
@@ -3,195 +3,88 @@
   width: 100%;
 }
 
-.md-select--medium {
-  max-width: 440px;
-}
-
-.md-select--small {
-  max-width: 260px;
-}
-
-.md-select__container {
-  position: relative;
-}
-
-.md-select__label {
-  display: flex;
-  align-items: flex-end;
-  font-weight: 600;
-}
-
+/* Label */
 .md-select__label-wrapper {
   margin-bottom: 0.5rem;
 }
-
-.md-select__label > * + * {
-  margin-left: 1rem;
+.md-select__label {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+.md-select__label label {
+  font-weight: 600;
 }
 
 .md-select__button {
-  font-family: 'Open sans';
-  font-size: 1rem;
-  width: 100%;
   display: flex;
   align-items: center;
   justify-content: space-between;
-  background-color: #fff;
-  padding: 1rem;
-  border: 1px solid var(--mdGreyColor60);
-  color: var(--mdGreyColor);
-  text-align: left;
-  cursor: pointer;
-}
-
-.md-select__button.md-select__button--error {
-  border-color: var(--mdErrorColor);
-}
-
-.md-select__button.md-select__button--small {
-  padding: 0.75rem;
-}
-
-.md-select__button:not(.md-select__button--open):focus.md-select__button--small {
-  padding: calc(0.75rem - 1px);
-}
-.md-select__button.md-select__button--open.md-select__button--small {
-  padding: calc(0.75rem - 1px);
-  padding-bottom: 0.75rem;
-}
-
-.md-select--disabled .md-select__button {
-  background-color: var(--mdGreyColor20);
-  color: var(--mdGreyColor60);
-  cursor: not-allowed;
-}
-
-.md-select__button:focus,
-.md-select__button:focus-within {
-  outline: none;
-}
-
-.md-select__button:not(.md-select__button--open):focus,
-.md-select__button:not(.md-select__button--open):focus-within {
-  padding: calc(1rem - 1px);
-  border: 2px solid var(--mdPrimaryColor);
-}
-
-.md-select--error.md-select__button:not(.md-select__button--open):focus,
-.md-select--error.md-select__button:not(.md-select__button--open):focus-within {
-  border-color: var(--mdErrorColor);
-}
-
-.md-select__button-text {
-  display: flex;
-  flex-grow: 1;
-  padding-right: 1rem;
-  min-height: 1.375rem;
-}
-
-.md-select__button-icon {
-  display: flex;
-  flex-shrink: 0;
-  width: 1.25rem;
-  height: 1.25rem;
-  rotate: 90deg;
-  color: var(--mdGreyColor);
-}
-
-.md-select__help-text {
-  max-height: 0;
-  overflow: hidden;
-  transition: max-height 0.15s ease-out;
-}
-
-.md-select__help-text--open {
-  padding-top: 0.5rem;
-  max-height: 2000px;
-  transition: max-height 0.5s ease-in;
-}
-
-.md-select__dropdown {
-  position: absolute;
-  max-height: 0;
-  overflow: hidden;
-  z-index: -1;
-  opacity: 0;
-  transition: max-height 0.5s ease-in-out;
   width: 100%;
+  font-size: 1rem;
+  line-height: 150%;
+  border: 1px solid var(--mdPrimaryColor);
+  padding: 0.75rem 1rem;
+  background-color: #fff;
 }
-
-.md-select__dropdown-item {
+.md-select--large .md-select__button {
+  padding: 1rem;
+}
+.md-select--small .md-select__button {
+  padding: 0.5rem 1rem;
+}
+.md-select__button::placeholder {
+  color: var(--mdGreyColor60);
+  font-size: 1rem;
+}
+.md-select__button[data-focus-visible],
+.md-select__button[aria-expanded='true'] {
+  outline: 2px solid var(--mdPrimaryColor);
+  outline-offset: -2px;
+}
+.md-select__button[aria-expanded='true']::placeholder {
+  opacity: 0;
+}
+.md-select__button:disabled {
+  background-color: var(--mdGreyColor20);
+  border-color: var(--mdGreyColor60);
+}
+.md-select__button:disabled::placeholder {
+  color: var(--mdGreyColor60);
+}
+.md-select__button .md-select__button-icon {
+  width: 1.5rem;
+  height: 1.5rem;
+  rotate: 0deg;
+  transition: rotate 0.2s ease-in;
+}
+.md-select__button[aria-expanded='true'] .md-select__button-icon {
+  rotate: 180deg;
+  transition: rotate 0.2s ease-in;
+}
+.md-select__button-right {
   display: flex;
   align-items: center;
-  font-family: 'Open sans';
-  border: 0;
+  gap: 0.5rem;
+}
+
+/* popover */
+.md-select__popover {
   background-color: #fff;
-  font-size: 1rem;
-  width: 100%;
-  text-align: left;
-  padding: 0.9rem;
-  transition: background-color 0.15s ease-in-out;
-  cursor: pointer;
-}
-
-.md-select__dropdown-item:hover,
-.md-select__dropdown-item:focus {
-  outline: none;
-  background-color: var(--mdPrimaryColor20);
-  transition: background-color 0.15s ease-in-out;
-}
-
-.md-select__dropdown-item--selected {
-  background-color: var(--mdPrimaryColor10);
-}
-
-.md-select__dropdown-item-text {
-  display: flex;
-  flex-grow: 1;
-}
-
-.md-select__dropdown-item-clear {
-  display: flex;
-  flex-shrink: 0;
-  height: 1.25rem;
-  width: 1.25rem;
-  color: var(--mdPrimaryColor);
-}
-
-/* Open state */
-.md-select--open .md-select__button {
-  border-top: 2px solid var(--mdPrimaryColor);
-  border-right: 2px solid var(--mdPrimaryColor);
-  border-left: 2px solid var(--mdPrimaryColor);
-}
-
-.md-select__button--open {
-  padding: calc(1rem - 1px);
-  padding-bottom: 1rem;
-}
-
-.md-select--open .md-select__dropdown {
-  max-height: 350px;
-  overflow-y: auto;
-  opacity: 1;
-  transition: max-height 0.5s ease-in-out;
+  max-height: min(var(--popover-available-height, 300px), 300px);
+  overflow: auto;
+  overscroll-behavior: contain;
   border: 2px solid var(--mdPrimaryColor);
-  box-shadow: 0px 2px 4px 0px rgba(0, 0, 0, 0.25);
   border-top: 0;
-  z-index: 2;
+  opacity: 0;
+  transition-duration: 200ms;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  animation-duration: 200ms;
+  animation-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transform: translateY(-5%);
+  z-index: 3;
 }
-
-/* open + error */
-.md-select--open.md-select--error .md-select__button {
-  border-color: var(--mdErrorColor);
-}
-.md-select--open.md-select--error .md-select__dropdown {
-  border-color: var(--mdErrorColor);
-}
-
-/* Error text */
-.md-select__error {
-  color: var(--mdErrorColor);
-  font-size: 0.88rem;
-  margin-top: 0.5rem;
+.md-select__popover[data-enter] {
+  opacity: 1;
+  transform: translateY(0%);
 }
